requirements.txt
Package             Version  
------------------- ---------
asgiref             3.3.1    
certifi             2020.11.8
cffi                1.14.3   
chardet             3.0.4    
cryptography        3.2.1    
defusedxml          0.6.0    
Django              3.1.3    
django-allauth      0.43.0   
django-cors-headers 3.7.0    
django-crispy-forms 1.9.2    
django-filter       2.4.0    
django-rest-auth    0.9.5    
djangorestframework 3.12.2   
Faker               5.0.1    
idna                2.10     
Markdown            3.3.3    
oauthlib            3.1.0    
Pillow              8.0.1    
pip                 19.2.3   
pycparser           2.20     
PyJWT               1.7.1    
python-dateutil     2.8.1    
python3-openid      3.2.0    
pytz                2020.4   
requests            2.25.0   
requests-oauthlib   1.3.0    
setuptools          41.2.0   
six                 1.15.0   
sqlparse            0.4.1    
text-unidecode      1.3      
urllib3             1.26.2   
